Add logic to disable TLS auth
By allowing self signed certificates the mediator can easily be setup
and tested in a dev environmnet.
Disabling TLS auuth is insecure therefore there is a warning logged and
it has to be explicitly set but this mediator shouldn't be used in any
roduction environment so there isn't ay risk.

OHM-796