Fix server port number
Copy paste error. The port specified in the constants file did not match
that set in the mediator config.

Perhaps this is an issue as the Server Port is a configurable variable
whilst the mediator Config is not.

Perhaps the mediator config needs to be created from the constants as
this would include the environment variable configuration...

OHM-796